requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-model-utils==4.5.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing-extensions==4.11.0